requirements.txt
fastapi
uvicorn
python-multipart
pillow
pdf2docx
python-docx
firebase-admin
python-dotenv